+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ability to display downloaded images
 - special for image displaying theme
 - multiple selection mode in BaseSelect
+- BaseColorPicker with more color choices, transparency changes and color model changes
 - label displaying in BaseRadio 
 - support 'max' and 'min' attributes for BaseInput
 - ability to change placement in BaseRadio